requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
pydantic-settings
openai>=1.30.0
tiktoken
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
chromadb>=0.5.0
